[feat] 添加订单详情【申购、赎回、转存】接口

+++ src/main/java/com/tianli/financial/entity/FinancialRecord.java
@@ -69,7 +69,7 @@ public class FinancialRecord {
     /**
      * 参考年化
      */
-    private double rate;
+    private BigDecimal rate;
 
     /**
      * 申购时间
